+++ NEWS.md
@@ -4,9 +4,19 @@ NEWS
 
 <!-- NEWS.md is generated from NEWS.Rmd. Please edit that file -->
 
-    #> Last Update: 2023-12-14 17:00:50.380076
+    #> Last Update: 2024-01-09 17:47:25.602262
 
 # Version History
+
+## Changes in version 1.0.0.9059 (2024-01-09)
+
+- feat: Add new issue template for metric and index requests, Issue \#99
+- refactor: Add metrics and index for MN IBI (bugs and fish), Issue \#92
+  - MetricNames.xlsx
+  - MetricScoring.xlsx
+  - `metric.values()`
+- tests: Run tests for new metrics
+  - Tweak data and tests as needed
 
 ## Changes in version 1.0.0.9058 (2023-12-14)